+++ src/aioway/_common/_typing.py
