+++ src/aioway/_keyed.py
@@ -3,28 +3,21 @@
 "The module containing `Fate` interface, the implementation for fake aten operations."
 
 import abc
-import dataclasses as dcls
 import inspect
+import re
 import typing
 from collections import abc as cabc
 
-from aioway._common import dcls_no_repr, render_fcall
-
-__all__ = ["Op"]
+__all__ = ["Keyed"]
 
 
-@dcls_no_repr
-class Op[K: cabc.Callable[..., object]](abc.ABC):
+_CAMEL_CASE_REGEX = re.compile(r"(?<!^)(?=[A-Z])")
+
+
+class Keyed[K: cabc.Callable[..., object]](abc.ABC):
     """
-    `Op` stands for [o]verridable [p]ass. Or [op]erator. It follows a pattern:
-    it sits in a family of similar operations, and can be looked up by a key (and signature).
-
-    An op is a custom (`aioway`) operation that may override the operator's behaviors at runtime,
-    or at least provide some static, inspectable info on the current call.
-
-    Right now, there are 2 `Op` kinds:
-    1. `Fate` for ATen operations.
-    2. `Might` for `nn.Module` init.
+    `Keyed` is a class where subclasses are indexed by their `KEY` attributes,
+    providing easy way to find the subclass by walking the subclass tree.
     """
 
     KEY: typing.ClassVar[K] = NotImplemented
@@ -45,29 +38,6 @@ class Op[K: cabc.Callable[..., object]](abc.ABC):
 
         if not callable(cls.KEY):
             raise TypeError(f"Non callable function key for {cls=}, {cls.KEY=}.")
-
-    @typing.override
-    def __repr__(self) -> str:
-        return render_fcall(f"{self.name()}", **dcls.asdict(self))
-
-    @typing.override
-    def __hash__(self) -> int:
-        return id(self)
-
-    @abc.abstractmethod
-    def do(self) -> typing.Any:
-        """
-        Generate the fake tensor.
-        """
-
-        raise NotImplementedError
-
-    @classmethod
-    @abc.abstractmethod
-    def name(cls) -> str:
-        "The name of the class to be rendered."
-
-        raise NotImplementedError
 
     @classmethod
     def find(cls, key: K) -> cabc.Generator[type[typing.Self]]:
@@ -101,12 +71,20 @@ class Op[K: cabc.Callable[..., object]](abc.ABC):
         yield from _iter_ops(cls)
 
     @classmethod
+    def _name(cls):
+        return _camel_to_snake(cls.__qualname__)
+
+    @classmethod
     def is_concrete(cls) -> bool:
         # Concrete in class var and concrete in methods.
         return cls.KEY is not NotImplemented and not inspect.isabstract(cls)
 
 
-def _iter_ops[T: Op](cls: type[T]) -> cabc.Generator[type[T]]:
+def _camel_to_snake(name: str) -> str:
+    return re.sub(_CAMEL_CASE_REGEX, "_", name).lower()
+
+
+def _iter_ops[T: Keyed](cls: type[T]) -> cabc.Generator[type[T]]:
     for sub in cls.__subclasses__():
         if sub.is_concrete():
             yield sub
